NEMPUB-660: Definition change for eTimes.13
Correcting a type in the old definition of eTimes.13

Old Value: The date/time the unit back was back in service and available for response (finished with call, but not necessarily back in home location).

New Value : The date/time the unit was back in service and available for response (finished with call, but not necessarily back in home location).